+++ dsr_agents/CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog for package dsr_agents
 * Fix get_zones in navigation node.
 * Battery agent will publish battery_percentage between 0 and 100.
 * Cancel goal before sending a new one.
+* Convert the power supply status from enum to string.
 
 0.5.0 (07-11-2023)
 ------------------